actualizacion

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
@@ -14,6 +15,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        Gate::authorize('read_user');
         return view('admin.users.index');
     }
 
@@ -22,8 +24,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create_user');
         $roles = Role::all();
-
         return view('admin.users.create', compact('roles'));
     }
 
@@ -32,6 +34,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create_user');
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users',
@@ -72,6 +75,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        Gate::authorize('read_user');
         return view('admin.users.show', compact('user'));
     }
 
@@ -80,6 +84,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        Gate::authorize('update_user');
         $roles = Role::all();
         return view('admin.users.edit', compact('user', 'roles'));
     }
@@ -89,6 +94,7 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        Gate::authorize('update_user');
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users,email,' . $user->id,
@@ -121,6 +127,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        Gate::authorize('delete_user');
         $user->roles()->detach();
         $user->delete();
